Adds tests for RailFence Cipher & Renames RailFence.js to RailFenceCipher.js
