+++ src/Encoder/RailFenceCipher.js
